+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## March 18th 
 
+- Refactored install.sh
+- Fixed file endings
 - Enabled support for .env files on Customfile
 - Added support to update vagrant box to the setup
 - Upgraded files to setup PHP 7.4